Add a reply to secondary color in process
Adds secondary color as input option well as a function returning an answer
in case of secondary color.

Furthermore the process function now is one big try - except block.
The expectation handler should be made to throw an exception when an
non-present expectation is being removed instead of printing to the console.